+++ apps/content-blocks-sandbox/src/Controller/PageController.php
@@ -7,7 +7,7 @@ namespace App\Controller;
 use App\Entity\Page;
 use ContentBlocks\Entity\ContentArea;
 use ContentBlocks\Entity\Section;
-use ContentBlocks\Service\SectionClonerInterface;
+use ContentBlocks\Section\SectionClonerInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;