+++ packages/content-blocks/tests/Publishing/ContentAreaPublisherTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace ContentBlocks\Tests\Service;
+namespace ContentBlocks\Tests\Publishing;
 
 use ContentBlocks\Entity\Block;
 use ContentBlocks\Entity\Column;
 use ContentBlocks\Entity\ContentArea;
 use ContentBlocks\Entity\Section;
-use ContentBlocks\Service\ContentAreaPublisher;
+use ContentBlocks\Publishing\ContentAreaPublisher;
 use Doctrine\ORM\EntityManagerInterface;
 use PHPUnit\Framework\TestCase;